クラスファイル追加
