+++ class/branch.php
@@ -7,7 +7,7 @@ class Branch
 		{
 			$id = $_POST['edit'];
 			$id = htmlspecialchars($id,ENT_QUOTES,'UTF-8');
-			header('Location: create_and_edit.php?edit='.$id);
+			header('Location: ../create_and_edit.php?edit='.$id);
 			exit();
 		}
 		
@@ -15,7 +15,7 @@ class Branch
 		{
 			$id = $_POST['delete'];
 			$id = htmlspecialchars($id,ENT_QUOTES,'UTF-8');
-			header('Location: delete_check.php?delete='.$id);
+			header('Location: ../delete_check.php?delete='.$id);
 			//header('Location: delete.php?delete='.$id);
 			exit();
 		}
@@ -24,7 +24,7 @@ class Branch
 		{
 			$search = $_POST['search_word'];
 			//$search = htmlspecialchars($search,ENT_QUOTES,'UTF-8');
-			header('Location: search_result_display.php?search='.$search);
+			header('Location: ../search_result_display.php?search='.$search);
 			//header('Location: index_search_display.php?search='.$search);
 		}
 	}
