+++ index.php
@@ -1,13 +1,11 @@
 <?php
-//4/20夜にindex2.phpから移植してきた。それに伴いpage.phpも新規で作成。
-//4/22 SQL文を記述するメソッドはmysql.php/MySQLクラスへ移動
+require_once ("class/indexclass.php");
+//require_once ("class/mysql.php");
+require_once ("class/page.php");
 
-require_once ("mysql.php");
-require_once ("page.php");
-
-$index = new MySQL();
+$index = new Index();
 $page  =new Page();
-$totalCount = $page->dataCount();
+$totalCount = $index->dataCount();
 $totalPage = $page->totalPage($totalCount);
 $nowPage = $page->nowPage($totalPage);
 $results = $index->get5IndexData($nowPage);
