+++ search_result_display.php
@@ -1,20 +1,5 @@
-<?php 
-//require_once "index.php"; 
-require_once "search_result.php";
-//require_once "page.php";
-?>
-
-<!doctype html>
-<html lang="ja">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <title>Index Page</title>
-</head>
+<?php include 'inc/head.php'; ?>
+<?php require_once "search_result.php"; ?>
 <body class="pl-4 pr-4">
 <header>
     <div class="mt-5 mb-3">
@@ -23,7 +8,7 @@ require_once "search_result.php";
 </header>
 <div class="">
     <!--検索結果表示-->
-    <p>検索ワード：<?php echo $search; ?><br /></p>
+    <p>検索ワード：<?php echo $search_word; ?><br /></p>
     <p>ヒット件数：<?php echo $totalCount; ?><br /><br /></p>
 </div>
 <!--ToDoリスト表示-->
@@ -41,7 +26,7 @@ require_once "search_result.php";
         </tr>
         <?php foreach($results as $result): ?>
             <?php if($result == false) break;?>
-            <form method="post" action="branch.php">
+            <form method="post" action="class/branch.php">
                 <tr>
                     <td><?php echo $result['id'];?></td>
                     <td><?php echo $result['title'];?></td>
@@ -56,14 +41,9 @@ require_once "search_result.php";
     </table>
 </div>
 <div class="btn btn-outline-secondary">
-    <?php $searchPage->pageSearchDisplay($nowPage, $totalPage, $previewNext, $search); ?>
+    <?php $searchPage->pageSearchDisplay($nowPage, $totalPage, $previewNext, $search_word); ?>
 </div>
 <a href='./index_display.php' style='padding: 5px;'>ToDo一覧</a>
-
-<!-- Optional JavaScript -->
-<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+<?php include 'inc/bootstrap.php';?>
 </body>
 </html>
